[Profiles] Add keep alive for profile statistics when deleting a profile
This CL creates a kProfileStatistics keep alive used for displaying the
profile statistics when deleting a profile from the Profile Picker.

The keep alive is released if the user cancels the profile deletion
dialog or closes the Picker, unloading the profile if that was the only
keep alive and the profile is not needed in memory anymore, which was
not the case before.

The keep alive is also removed if the user continues with the profile
deletion.

Preview:
https://screencast.googleplex.com/cast/NjA3ODAyOTMwNzU3NjMyMHw3Y2M5NTMzOS0wMw

Fixed: 1434951
Change-Id: Ied6d40bddf3375c9ce32bf392174ee230e59740c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4474512
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Gabriel Oliveira <gabolvr@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135911}