Avoid dangling ptr warning in DeviceTrustServiceTest
Ownership of the DeviceTrustService is currently passed to the test
body after creation. However, there are unowned raw_ptr<> members in
the test class to its owned sub-resources that outlive it. Instead,
keep ownership of the service in the test class, passing an unowned
reference to the test body.

Bug: 1401495
Change-Id: Ia558d5aa09aaf908c8f06da5d7dcc6dd650f3fe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4476084
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sebastien Lalancette <seblalancette@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135918}