[Read Anything] Fix 'Open in reading mode' context menu item on mac
This CL fixes a bug where the "Open in reading mode" context menu item
does not open the side panel on Mac. Previously, we opened the side
panel in render_view_context_menu_views.cc but that file is only used
in some builds and not in Mac builds.

To address this bug, this CL moves the call to open the side panel to
render_view_context_menu.cc, which is included in Mac builds.

Also, we add a check to make sure the reading mode context menu item
only shows if reading mode is not already opened in the side panel.

Bug: 1266555, 1425968
Change-Id: Ie3acfb0c24151f1f3ed72265425ae6131052467e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4377302
Reviewed-by: Scott Violet <sky@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Jocelyn Tran <jocelyntran@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1135923}