[DataRetentionPolicies] Adding a Mapping of Policy Data Type Name to Int
BrowsingDataType that was used to check that all browsing data policy
types are mapped to sync types is actually for the types shown in the UI.
These do not always map to the policy types as we can have a browsing
data type to be controlled by policy without adding it to the UI, and
vice versa.

This CL adds an enum for browsing data types that can be managed by
policy. By using an enum and making the strings only visible to the utils
we can ensure that the strings are no longer used directly. If a new
type string is added, an enum integer has to  correspond to it and be
mapped to policy sync types. If this is not the case an assertion fails.

Bug: b/272025211
Change-Id: I6e1f9dcf66f11bceed6703733c5c20db424c9957
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4468469
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135948}