media/gpu/v4l2/test: Update H.264 CAPTURE queue buffer refresh
The H.264 decoder was previously setup to refresh CAPTURE queue buffers
based on which buffers had a different buffer ID than the current buffer
(as to not override the current data). This becomes an issues especially
if a picture is a long term reference and it's data gets lost.

This change updates DPB to indicate the CAPTURE queue indices of its
pictures which have not yet been output. The
GetReusuableReferenceSlots() function is updated as well to prevent
refreshing buffers which have a DPB element not yet output attached to
them.

This change results in all h264_baseline and h264_main bitstreams
passing their tests.

BUG=b:234752983
TEST=tast -verbose run dut video.PlatformDecoding.v4l2_stateless_h264_baseline
TEST=tast -verbose run dut
video.PlatformDecoding.v4l2_stateless_h264_main

Change-Id: I6352710eb1444bdfd1028462ca5d16f3ba5c0f73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4409895
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135954}