media/gpu/v4l2/test: Update H.264 Decoder logging messages
Previously the H.264 decoder was setup with logging messages which used
DVLOGF. Since a debug build is not usually used these would not work.

This update replaces the logging calls to DVLOGF to VLOGF with high
verbosity.

This has no effect on the decoder's decoding ability but instead on the
ability to see logging on a non debug build.

BUG=b:234752983
TEST=v4l2_stateless_decoder --video=BA1_Sony_D.h264 --md5=checksums.log
TEST=v4l2_stateless_decoder --video=CVSE2_Sony_B.h264 --md5=checksums.log

Change-Id: I10c693931980bbfc1f63a3c780747a48492a1443
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4475676
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135960}