Re-enable TabModelSelectorTabObserverTest tests
For some reason a change from January or earlier suddenly started
flaking the TabModelSelectorTabObserverTest.

What appears to have occurred is that deferred tab registrations from
the UI thread are racing with the test asserts which are happening on
the instrumentation thread. The easiest solution is to just wait for
the deferred operation to resolve.

In production this should never happen because all use of this observer
should take place on the UI thread.

Locally repeated 50 times with 0 flakes after this change it flaked
40% of the time before.

Bug: 1435130
Change-Id: I05b1d372dc0f7c78b4a4e4f3bc1f3c4d8b0df1a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4476188
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135966}