Add Ruler for Java at 100 characters
The Chromium Java style guide states nothing about the required line
length, which implies that the line length is mandated by the Android
style guide, which states that it should not be more than 100
characters:
https://source.android.com/docs/setup/contribute/code-style#limit-line-length
This is also backed up by the behavior of clang-format.

This change adds a language-specific ruler override for Java to change
the java ruler from the default length of 80 to a length of 100.

Change-Id: I959b3a00f417374fe89a8db9e99950c9133d5b8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4477692
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1135997}