[Performance Controls] Reland introduce discard time pref
This is a reland of crrev.com/c/4418953, which was reverted due to
causing crashes (crbug.com/1436044). The issue was SetTimeBeforeDiscard
has a CHECK that HighEfficiencyModePolicy exists, so when the heuristic
memory saver flag was on, HighEfficiencyModePolicy is not created and
so the check fails. The HighEfficiencyModePolicy is not needed when the
flag is on, so the fix is just to check that the flag is off before
attempting to set the discard time. A test is added to verify that the
UPTM starts correctly when the feature flag is on.

Original CL description:
Creates a pref to control the discard time in minutes for memory saver.
The UI will be implemented at a later date, so to control the pref this
CL also introduces a command line switch that the UPTM reads to set the
default value of the pref. Once the multistate memory saver is rolled
out, the user will be able to control the pref from the settings page,
and this command line switch will no longer be needed.

Bug: 1424220
Change-Id: Iee1a848f725747f7633d25c12786ba5b8ed00388
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4475131
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136008}