HTTPS-First Mode: Enable Site Engagement heuristic with metrics
This CL implements Site Engagement Heuristic to enable HTTPS-First Mode
on sites that the user interacts with regularly. If the engagement
score of the HTTPS URL of a host goes above a certain threshold, HFM
is enabled on that host. Presently, this CL only records a new histogram
with the same events as HFM.

The current thresholds for engagement scores are arbitrary and will be
tweaked in future CLs. This CL also only enables HFM for a site, but
doesn't remove it. This will be implemented in future CLs.

Bug: 1435222
Change-Id: I790a2de8506282a265ad7154d50a394f90227618
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4459815
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136015}