[Auto-disable Accessibility] Update AccessibilityState system observers and init timeline
This CL continues the AccessibilityState refactoring/improvements as
part of the auto-disable accessibility feature.

This CL updates the AccessibilityState class to register as a listener
of the ActivityState in ApplicationStatus. A user can change the
enabled/disabled accessibility services while Chrome is in the
background, so when an Activity is resumed, AccessibilityState should
query the services again. We also add an unregister method when
all Activities have been destroyed.

We also move the ContentObserver registration out of the
browser_accessibility_state_impl_android, and into the Java-side
init process methods. The AccessibilityState observations should
not be tied to the web contents, and should exist even when there is
no webContents or tab so the UI can use the same state machine.
During postNativeInit we register the system observers, and in
startup add a deferred task to query the state so that it can
be ready for clients asap.

We also move the collection of AccessibilityServiceInfo histograms
into the AccessibilityState init process, and any time the system
settings change. This is cleaner than the current method which
records the stats anytime a new webContents object is created,
since such stats are not inherently tied to webContents.

We add some asserts to ensure that AccessibilityState system
observer registration happens before any client queries of state.


AX-Relnotes: N/A
Bug: 1362100
Change-Id: I03272d426ca5a7c719a2c01d522de3b495175dcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4313105
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Cr-Commit-Position: refs/heads/main@{#1136037}