[Tab Switcher] Skip Toolbar Visibility Animation for Touch Exploration
https://chromium-review.googlesource.com/c/chromium/src/+/4456310 fixed
the animation janking by making the tab to grid animation instant, but
the toolbar still remains visible for a short period after the
transition finishes as it is part of a separate animation.

This CL also disables the toolbar visibility animation for both the
cases where Start Surface is and isn't enabled.

This only applies when Touch Exploration is enabled as that is what
causes focus issues during the animation.

Fix videos prefixed with part2:
https://drive.google.com/drive/folders/1xnki8Emrlvo1n1TuPZHYgXMbTb9ssYBz?usp=sharing

Bug: 1434612
Change-Id: I8433598ac3f853832cffe1dda02e25b9971a124f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4466968
Reviewed-by: Xi Han <hanxi@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136040}