iwa: Make sure that IWA installation from CLI is working on ChromeOS
Our intention was that `MaybeInstallIwaFromCommandLine` is called on
browser startup. However, the browser's complex startup behavior in
conjunction with the many different ways it is started on ChromeOS (Ash
vs. Lacros, first run vs. not first run, etc.) made it so that
`MaybeInstallIwaFromCommandLine` is not called consistently. Instead of
trying to catch all the different startup behaviors of the browser on
ChromeOS, I instead moved the installation logic into
`IsolatedWebAppCommandLineInstallManager::Start` for ChromeOS.

This is something that should be revisited once we have proper developer
UI to install IWAs, and once Lacros becomes the default.

Bug: 1426677, b/272471661, b/274758125
Change-Id: I241c937f036447df6ef2b55e15ba9f5a920d0c7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4466057
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136041}