Add new variants to patterned ambient mode histograms.
This is to account for the new video screensaver that's in
the process of being developed. There are several patterned
histograms that track some metric for each set of screensaver
"settings" (slide show, lottie animation, and now video). The
semantics of these histograms remain exactly the same as before,
and we still record them for existing screensaver settings that
are already available in the field. These video metrics will
be implemented for existing histograms in the next CL.

Bug: b:271182121
Change-Id: Ibb5a5b20d83964a79ce1f5f229a1252fd822cf39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4477584
Reviewed-by: Yulun Wu <yulunwu@chromium.org>
Commit-Queue: Eric Sum <esum@google.com>
Cr-Commit-Position: refs/heads/main@{#1136048}