[FreezeCast] Fix DCHECK failures
cast_mirroring_service_host now retains a reference to
video_capture_host_, a weak_ptr. We were holding on to this reference
for too long in a few cases, which led to DCHECK failures. This change
ensures that we drop the reference in the cases where
video_capture_host_ will be invalidated.

Also, ensure that the UI is notified in that case that mirroring
resumes because the source has switched (tab switching).

Bug: b:279450754, b:279450081, b:279449269
Change-Id: I0b82c6e88437a9d318d03e844ab7066392ad4198
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4471676
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Benjamin Zielinski <bzielinski@google.com>
Cr-Commit-Position: refs/heads/main@{#1136218}