Only run predicate if predicate is not null
In StandaloneBrowserExtensionAppShelfItemController::ItemSelected(), the
filter_predicate passed in as its fifth argument sometimes can be
base::NullCallback(). In this case, the predicate should be skipped and
should not be run. Otherwise, DCHECK like this [1] in scoped_refptr
would be triggered.

[1] https://source.chromium.org/chromium/chromium/src/+/refs/tags/114.0.5731.1:base/memory/scoped_refptr.h;l=286;bpv=1;bpt=0

Bug: 1436051
Test: manual verification + CQ
Change-Id: Iaad727f4439be4ecdcec3f834607c5ccb05db52e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4466368
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Di Wu <diwux@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136257}