media/cast/encoding: Enable h264 hardware encoder on ChromeOS AMD
This CL enables h264 hardware encoder. It is disabled because the
connection to chromecast fails with the encoder probably due to the
decoder failure on chromecast. The chromecast device is probably third
generation chromecast, but I don't have it. We will attempt enabling
and see lab test results.

Bug: b:178660032
Test: Cast tab on grunt to ChromeCast with Google TV and ChromeCast 2nd gen

Change-Id: Ib0321d75937b6ec6fcea4d7c93648c2b914ccd9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4483074
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136300}