[FilesBulkPinning] Create container for the cloud panel
The store maintains the state for the bulk pinning feature and the
<xf-cloud-panel> maintains the visual representation via the cloud icon.
To map the state to the element, introduce the CloudPanelContainer which
takes the data from the store and updates the element appropriately.
This currently just updates in progress state, further CLs will update
the error states and paused states.

Bug: b:262647382
Test: browser_tests --gtest_filter=--gtest_filter=*CloudPanelContainer
Change-Id: I6c168c3b977422b9c6f7c6dbbb027058df7fd23f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4478092
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136311}