Update ScreenAIServiceRouter to pause connections till library is ready.
ScreenAIServiceRouter may receive connections before library is loaded
and initialized and ScreenAI service had to deprioritize all requeusts
until library is ready.
After this change, the connections are queued until the library is ready
and the service can be simplified (in a next CL) to run the requests as
they come.

AX-Relnotes: n/a.
Bug: 1278249
Change-Id: If8b0059e3ff8de80758a2a2122cd638104df0512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4479411
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136336}