Allow history deletion for supervised users in CBD on Desktop platforms
This updates the logic in the Clear Browsing Data dialog on Desktop
platforms so that it is allowed for supervised users, if they have
history deletion allowed in other UI surfaces.

Note that this does not change the behaviour on ChromeOS, until the
AllowHistoryDeletionForChildAccounts flag is enabled.

Bug: b/251192695
Change-Id: If28da69c069d2774651f79eb29f17eb3a186da99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4442255
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: James Lee <ljjlee@google.com>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1136373}