[BSC]: Implement `BoundSessionRequestThrottledInRendererManager`
`BoundSessionRequestThrottledInRendererManager` is responsible for
notifying the browser that there are one or more requests blocked in the
renderer as it sees the required short lived cookie as expired.

On a follow-up CL, `BoundSessionRequestThrottledInRendererManager` will
be wrapped  in `BoundSessionRequestThrottledListener` and passed to
`GoogleURLLoaderThrottler`. This allows the renderer to request a
refresh from the browser and allow it to resume requests on success ,
timeout or failure.

There is a single mojo pipe from the browser process to a renderer
process. `BoundSessionRequestThrottledListener` owns this pipe.
`GoogleURLLoaderThrottler` has one instance per network request and can
be created on different threads. Therefore calls to
`BoundSessionRequestThrottledListener` must be done one a single thread
`IO` thread.

Bug: b/263264391, 1392135
Change-Id: Ic1ad8752502cbddfea0086cfffbbc61864ff5101
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4474292
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136382}