Remove comment about making midi sysex syncable
The MIDI SysEx permission was initially syncable, and was made
unsyncable when permission delegation was launched in 2018.  There is a
TODO to make it syncable again.  However, the most recent guidance is to
make permissions unsyncable unless there is a strong reason to make
them syncable.  Since there is no strong reason to make this permission
syncable, remove the TODO comment and leave it as-is.

Bug: 879954
Change-Id: I4cdb998a7eb9c06be95f4f184390f682ba104052
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4481112
Auto-Submit: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136404}