Replace manual fetch/XHR Attribution-Reporting headers with options
https://github.com/WICG/attribution-reporting-api/pull/757
https://github.com/WICG/attribution-reporting-api/pull/773

Previously, users of fetch and XHR opted in to Attribution Reporting by
manually setting the corresponding header. However, this was error-prone
and not precise, as the header could be set even when the browser would
reject registrations in the subsequent response, and required full
CORS support on the server. This change doesn't fully remove the CORS
requirements as a result of the responses still being processed in the
renderer.

We implement this by propagating the eligibility as a field on
ResourceRequest. The IDL, Mojo, and fetch/XHR changes are largely based
on the similar setup for Trust Tokens.

This also allow us to remove all related structured-header parsing
from //components/attribution_reporting, as the new ResourceRequest
field can be directly and far more efficiently inspected.

Moving this logic out of blink and content and into the network
service will make it easier to move all Attribution Reporting
response-handling there in the future, which is needed to address
crbug.com/1374121.

Bug: 1425083, 1374121, 1434431
Change-Id: Iab65e2c42435176b42d500ac252777195934689f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4419147
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Quick-Run: Andrew Paseltiner <apaseltiner@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136461}