Add dedicated cache entry object for shopping service
This patch replaces the tuple used in the product info cache with a
dedicated object. This makes the code considerably easier to reason
about and allows for easier addition of new items that track the state
of the cached info.

Change-Id: I8a0a317d57ad900408d75a310521d2eb840c3cad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4460031
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Zhiyuan Cai <zhiyuancai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136558}