[dPWA] Testing placeholder and install url edge cases.
This adds a multitude of tests for the Externally Managed App Manager
around edge cases with placeholders, install urls, and various types of
conflicts and funny configurations.

These tests document the current behavior - some of this may change
in the future, but we want to have the tests here to ensure that
whatever behavior we decide, there are tests for these strange
scenarios.

Bug: 1434692
Change-Id: I636b63dd4b45b5df1b9edaba194981b70ff70822
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4457274
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Auto-Submit: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1136569}